Fix tests

+++ scripts/analytics/op_profiling/s5/plot_op_s5.go
@@ -319,7 +319,7 @@ func main() {
 	for opId, count := range countByOpId {
 		if count > 0 {
 			opIds = append(opIds, opId)
-			opNameByOpId[opId] = tracer.OpText[uint16(opId)]
+			opNameByOpId[opId] = tracer.OpText[uint8(opId)]
 		}
 	}
